finish backend product and related features

+++ app/Http/Controllers/Backend/ProductVariantController.php
@@ -6,6 +6,7 @@ use App\DataTables\ProductVariantDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\Variant;
+use App\Models\VariantItem;
 use Illuminate\Http\Request;
 
 class ProductVariantController extends Controller
@@ -92,6 +93,13 @@ class ProductVariantController extends Controller
     public function destroy(string $id)
     {
         $variant = Variant::find($id);
+        $VariantItemCheck = VariantItem::where('product_variant_id', $id)->count();
+        if ($VariantItemCheck > 0) {
+            return response([
+                'status' => "error",
+                'message' => 'This Variant contains VariantItem. Please delete all variant item related this one'
+            ]);
+        }
         $variant->delete();
         return response([
             'status' => "success",
